feat(RHINENG-10392): migrate compliance systems page and SystemsTable component to REST
