+++ src/SmartComponents/ComplianceSystems/ComplianceSystems.js
@@ -8,6 +8,7 @@ import PageHeader, {
 import { StateViewPart, StateViewWithError } from 'PresentationalComponents';
 import { SystemsTable } from 'SmartComponents';
 import * as Columns from '../SystemsTable/Columns';
+import useAPIV2FeatureFlag from '../../Utilities/hooks/useAPIV2FeatureFlag';
 
 export const QUERY = gql`
   {
@@ -23,12 +24,29 @@ export const QUERY = gql`
   }
 `;
 
-const DEFAULT_FILTER = 'has_test_results = true or has_policy = true';
+const DEFAULT_FILTER_GRAPHQL = 'has_test_results = true or has_policy = true';
+const DEFAULT_FILTER_REST = 'assigned_or_scanned=true';
+
+const dataMap = {
+  display_name: 'name',
+  culled_timestamp: 'culledTimestamp',
+  os_major_version: 'osMajorVersion',
+  os_minor_version: 'osMinorVersion',
+  stale_timestamp: 'staleTimestamp',
+  stale_warning_timestamp: 'staleWarningTimestamp',
+  'policies[0].title': 'policies[0].name',
+  groups: 'groups',
+  id: 'id',
+  insights_id: 'insights_id',
+  tags: 'tags',
+  updated: 'updated',
+};
 
 export const ComplianceSystems = () => {
   const { data, error, loading } = useQuery(QUERY);
   const policies = data?.profiles?.edges.map(({ node }) => node);
-
+  const apiV2Enabled = useAPIV2FeatureFlag();
+  console.log('debug: Hello compliance systems');
   return (
     <React.Fragment>
       <PageHeader className="page-header">
@@ -40,9 +58,12 @@ export const ComplianceSystems = () => {
             {policies && (
               <SystemsTable
                 columns={[
-                  Columns.customName({
-                    showLink: true,
-                  }),
+                  Columns.customName(
+                    {
+                      showLink: true,
+                    },
+                    { sortBy: apiV2Enabled ? ['display_name'] : ['name'] }
+                  ),
                   Columns.inventoryColumn('groups', {
                     requiresDefault: true,
                     sortBy: ['groups'],
@@ -55,7 +76,9 @@ export const ComplianceSystems = () => {
                     transforms: [nowrap],
                   }),
                 ]}
-                defaultFilter={DEFAULT_FILTER}
+                defaultFilter={
+                  apiV2Enabled ? DEFAULT_FILTER_REST : DEFAULT_FILTER_GRAPHQL
+                }
                 systemProps={{
                   isFullView: true,
                 }}
@@ -67,6 +90,8 @@ export const ComplianceSystems = () => {
                 remediationsEnabled={false}
                 policies={policies}
                 showGroupsFilter
+                dataMap={dataMap}
+                apiV2Enabled={apiV2Enabled}
               />
             )}
           </StateViewPart>
