+++ src/Utilities/hooks/api/useOperatingSystems.js
@@ -1,0 +1,7 @@
+import useQuery, { apiInstance } from '../useQuery';
+
+const useOperatingSystemsQuery = () => {
+  return useQuery(apiInstance.systems);
+};
+
+export default useOperatingSystemsQuery;
