+++ src/Utilities/hooks/api/useSystemsQuery.js
@@ -1,0 +1,7 @@
+import useQuery, { apiInstance } from '../useQuery';
+
+const useSystemsQuery = () => {
+  return useQuery(apiInstance.systems);
+};
+
+export default useSystemsQuery;
